feat: Add comprehensive ARCHITECTURE.md for Sovereign Firm v2.0
Enterprise-Grade Agentic AI Consultancy Platform architecture design:

Agent System:
- Dynamic agent lifecycle and instance structure
- 20+ specialist roles (Cloud Architect, SRE, K8s, DevOps, etc.)
- Skill system with YAML definitions
- Multi-agent coordination with file locking
- Task Dependency Graph (DAG) for parallel execution
- Structured Output Contracts (SOC) for determinism
- Agent CI pipeline with self-correction loops
- Embedded Git for version control

Industry Standards (Jan 2026):
- MCP (Model Context Protocol) adoption plan
- A2A (Agent-to-Agent) protocol awareness
- Plan-and-Execute pattern (2026 gold standard)
- Multi-agent orchestration patterns
- Comparison: LangGraph, CrewAI, AutoGen, Temporal

Security & Safety:
- HAP (Hate, Abuse, Profanity) filtering
- Multi-layer sandbox (WebContainers, gVisor, Firecracker)
- PII detection and secret scanning
- Tenant and execution isolation

Infrastructure:
- Tree-sitter for language-agnostic code analysis
- Universal language support (Go, Python, Rust, Java, etc.)
- Intelligence Layer (RAG with ChromaDB)
- Streaming/chunked generation
- Cost management with model routing

Also includes:
- Azure OpenAI LLM provider support
- QA agent for test generation
- Frontend PodConsole updates

+++ cmd/verify/main.go
@@ -75,7 +75,7 @@ func checkChroma() {
 	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
 	defer cancel()
 
-	_, err := c.SimilaritySearch(ctx, []float64{0.1, 0.2}, 1)
+	_, err := c.SimilaritySearch(ctx, []float32{0.1, 0.2}, 1)
 
 	if err != nil {
 		fmt.Printf("❌ Failed: %v\n", err)
