+++ pkg/sovereign/memory/memory.go
@@ -6,7 +6,7 @@ type Document struct {
 	ID       string                 `json:"id"`
 	Content  string                 `json:"content"`
 	Metadata map[string]interface{} `json:"metadata"`
-	Vector   []float64              `json:"vector,omitempty"` // Computed embedding
+	Vector   []float32              `json:"vector,omitempty"` // Computed embedding
 }
 
 // VectorStore abstraction for Air-Gapped RAG
@@ -15,5 +15,5 @@ type Store interface {
 	AddDocuments(ctx context.Context, docs []Document) error
 
 	// SimilaritySearch finds relevant docs
-	SimilaritySearch(ctx context.Context, vector []float64, limit int) ([]Document, error)
+	SimilaritySearch(ctx context.Context, vector []float32, limit int) ([]Document, error)
 }
